requirements.txt
flask
flask-cors
pillow
numpy
onnxruntime